source/requirements.txt
requests
PyGithub
tzdata
urllib3